feat(app): load new public talks list to app
