+++ src/v3/features/meeting_materials/public_talks/tableView/talk_row/index.tsx
@@ -1,10 +1,6 @@
 import { Box, TableCell, TableRow } from '@mui/material';
-import { TalkLocaleType } from '@definition/sources';
 import Typography from '@components/typography';
-
-type TalkRowType = {
-  talk: TalkLocaleType;
-};
+import { TalkRowType } from './index.types';
 
 const TalkRow = ({ talk }: TalkRowType) => {
   return (
