+++ src/v3/features/meeting_materials/public_talks/tableView/talk_row/index.types.ts
@@ -1,0 +1,3 @@
+import { PublicTalkType } from '@definition/public_talks';
+
+export type TalkRowType = { talk: PublicTalkType };
