+++ src/v3/features/meeting_materials/public_talks/tableView/talk_row/useTableView.tsx
