+++ src/v3/features/public_talks/components/listView/talk_row/index.types.ts
@@ -1,0 +1,3 @@
+import { PublicTalkType } from '@definition/public_talks';
+
+export type TalkRowType = { talk: PublicTalkType; isExpandAll: boolean };
