+++ src/v3/services/recoil/publicTalks.ts
@@ -1,18 +1,16 @@
 import { promiseGetRecoil, promiseSetRecoil } from 'recoil-outside';
-import { publicTalksSearchKeyState, publicTalksLocaleState, publicTalksState } from '@states/publicTalks';
-import { TalkLocaleType, TalkType } from '@definition/sources';
+import { publicTalksSearchKeyState, publicTalksState } from '@states/publicTalks';
+import { PublicTalkType } from '@definition/public_talks';
 
 export const setPublicTalksSearchKey = async (value: string) => {
   await promiseSetRecoil(publicTalksSearchKeyState, value);
 };
 
 export const publicTalksFind = async (talk_number: number) => {
-  const talks: TalkType[] = await promiseGetRecoil(publicTalksState);
+  const talks: PublicTalkType[] = await promiseGetRecoil(publicTalksState);
   return talks.find((record) => record.talk_number === talk_number);
 };
 
-export const publicTalksFindLocal = async (talk_number: number) => {
-  const talks: TalkLocaleType[] = await promiseGetRecoil(publicTalksLocaleState);
-  const found = talks.find((talk) => talk.talk_number === talk_number);
-  return found?.talk_title || '';
+export const setPublicTalks = async (value: PublicTalkType[]) => {
+  await promiseSetRecoil(publicTalksState, value);
 };
